moves GCC/Clang 128-bit int fixups to their own file
- includes it in fixed_point.h
