+++ include/fixed_point_utils.h
@@ -214,58 +214,4 @@ namespace std {
     };
 }
 
-#if defined(_GLIBCXX_USE_INT128)
-// std type specializations required to promote 128-bit integers to first-class integers under GCC/Clang
-namespace std {
-    // std::is_integral - related to https://llvm.org/bugs/show_bug.cgi?id=23156
-    template<>
-    struct is_integral<__int128> : std::true_type {
-    };
-
-    template<>
-    struct is_integral<unsigned __int128> : std::true_type {
-    };
-
-    // std::is_signed
-    template<>
-    struct is_signed<__int128> : std::true_type {
-    };
-
-    template<>
-    struct is_signed<unsigned __int128> : std::false_type {
-    };
-
-    // std::is_signed
-    template<>
-    struct is_unsigned<__int128> : std::false_type {
-    };
-
-    template<>
-    struct is_unsigned<unsigned __int128> : std::true_type {
-    };
-
-    // std::make_signed
-    template<>
-    struct make_signed<__int128> {
-        using type = __int128;
-    };
-
-    template<>
-    struct make_signed<unsigned __int128> {
-        using type = __int128;
-    };
-
-    // std::make_signed
-    template<>
-    struct make_unsigned<__int128> {
-        using type = unsigned __int128;
-    };
-
-    template<>
-    struct make_unsigned<unsigned __int128> {
-        using type = __int128;
-    };
-}
-#endif
-
 #endif	// defined(_SG14_FIXED_POINT_UTILS_H)
